pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "disaster-response-ops"
version = "1.0.0"
description = "A high-fidelity disaster response triage and resource allocation environment."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "openai",
    "websockets",
    "python-dotenv",
    "requests",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["app*", "server*"]